Updated RotaryEncoder
Added functionality to detect rotations of the encoder with a state machine.
This method of rotation detection does not require any debounce handling but is prone to some race condition bugs due to interrupts being used to detect change in values for the encode pins. It works well enough for the task at hand so we won't be spending any more time perfecting this encoder.